Move samples in Mapsui.Tests.Common to Mapsui.Samples.Common
