+++ Samples/Mapsui.Samples.Common/Maps/Tests/BitmapAtlasSample.cs
