+++ Samples/Mapsui.Samples.Common/Maps/Tests/BitmapSymbolInCollectionSample.cs
