+++ Samples/Mapsui.Samples.Common/Maps/Tests/BitmapSymbolSample.cs
