+++ Samples/Mapsui.Samples.Common/Maps/Tests/CalloutSample.cs
