+++ Samples/Mapsui.Samples.Common/Maps/Tests/EmptySample.cs
