+++ Samples/Mapsui.Samples.Common/Maps/Tests/GeometryCollectionTestSample.cs
