+++ Samples/Mapsui.Samples.Common/Maps/Tests/LabelSample.cs
