+++ Samples/Mapsui.Samples.Common/Maps/Tests/LineSample.cs
