+++ Samples/Mapsui.Samples.Common/Maps/Tests/ProjectionTestSample.cs
