+++ Samples/Mapsui.Samples.Common/Maps/Tests/ShapefileZoomSample.cs
