+++ Samples/Mapsui.Samples.Common/Maps/Tests/StackedLabelsTestSample.cs
