+++ Samples/Mapsui.Samples.Common/Maps/Tests/SvgSymbolSample.cs
