+++ Samples/Mapsui.Samples.Common/Maps/Tests/SymbolTypesSample.cs
