+++ Samples/Mapsui.Samples.Common/Maps/Tests/TouchPointSample.cs
