+++ Samples/Mapsui.Samples.Common/Maps/Tests/VectorStyleSample.cs
