+++ Samples/Mapsui.Samples.Common/Maps/Tests/WidgetsSample.cs
