+++ Samples/Mapsui.Samples.Common/Mapsui.Samples.Common.csproj
@@ -152,6 +152,24 @@
     <EmbeddedResource Include="Images\netherlands.jpg" />
     <EmbeddedResource Include="Images\osm-liberty.png" />
     <EmbeddedResource Include="Images\Pin.svg" />
+    <EmbeddedResource Include="Resources\Images\avion_silhouette.png" />
+    <EmbeddedResource Include="Resources\Images\checkered.png" />
+    <EmbeddedResource Include="Resources\Images\circle.png" />
+    <EmbeddedResource Include="Resources\Images\Ghostscript_Tiger.svg" />
+    <EmbeddedResource Include="Resources\Images\iconthatneedsoffset.png" />
+    <EmbeddedResource Include="Resources\Images\osm-liberty.png" />
+    <EmbeddedResource Include="Resources\Images\Pin.svg" />
+    <EmbeddedResource Include="Resources\SampleTiles\0_0_0.png" />
+    <EmbeddedResource Include="Resources\SampleTiles\1_0_0.png" />
+    <EmbeddedResource Include="Resources\SampleTiles\1_0_1.png" />
+    <EmbeddedResource Include="Resources\SampleTiles\1_1_0.png" />
+    <EmbeddedResource Include="Resources\SampleTiles\1_1_1.png" />
+    <EmbeddedResource Include="Resources\Shapefiles\test_file.dbf" />
+    <EmbeddedResource Include="Resources\Shapefiles\test_file.prj" />
+    <EmbeddedResource Include="Resources\Shapefiles\test_file.qmd" />
+    <EmbeddedResource Include="Resources\Shapefiles\test_file.shp" />
+    <EmbeddedResource Include="Resources\Shapefiles\test_file.shp.sidx" />
+    <EmbeddedResource Include="Resources\Shapefiles\test_file.shx" />
   </ItemGroup>
 
   <ItemGroup>
