+++ Samples/Mapsui.Samples.Common/Resources/Images/avion_silhouette.png
