+++ Samples/Mapsui.Samples.Common/Resources/Images/checkered.png
