+++ Samples/Mapsui.Samples.Common/Resources/Images/circle.png
