+++ Samples/Mapsui.Samples.Common/Resources/Images/iconthatneedsoffset.png
