+++ Samples/Mapsui.Samples.Common/Resources/Images/osm-liberty.png
