+++ Samples/Mapsui.Samples.Common/Resources/SampleTiles/0_0_0.png
