+++ Samples/Mapsui.Samples.Common/Resources/SampleTiles/1_0_0.png
