+++ Samples/Mapsui.Samples.Common/Resources/SampleTiles/1_0_1.png
