+++ Samples/Mapsui.Samples.Common/Resources/SampleTiles/1_1_0.png
