+++ Samples/Mapsui.Samples.Common/Resources/SampleTiles/1_1_1.png
