+++ Samples/Mapsui.Samples.Common/Resources/Shapefiles/test_file.shx
