+++ Samples/Mapsui.Samples.Common/TestTools/TestLayer.cs
