--- Tests/Mapsui.Tests.Common/GlobalSuppressions.cs
@@ -1,18 +1,0 @@
-﻿// This file is used by Code Analysis to maintain SuppressMessage
-// attributes that are applied to this project.
-// Project-level suppressions either have no target or are given
-// a specific target and scoped to a namespace, type, member, etc.
-
-using System.Diagnostics.CodeAnalysis;
-
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.BitmapAtlasSample.CreateMapAsync~Mapsui.Map")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.BitmapSymbolInCollectionSample.CreateMapAsync~Mapsui.Map")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.BitmapSymbolSample.CreateMapAsync~Mapsui.Map")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.BitmapSymbolWithRotationAndOffsetSample.CreateMapAsync~Mapsui.Map")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.LineSample.CreateMapAsync~Mapsui.Map")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP003:Dispose previous before re-assigning.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.LineSample.CreateLineProvider~Mapsui.Providers.MemoryProvider{Mapsui.IFeature}")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP003:Dispose previous before re-assigning.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.PolygonTestSample.CreatePolygonProvider~Mapsui.Providers.MemoryProvider{Mapsui.IFeature}")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.PolygonTestSample.CreateMapAsync~Mapsui.Map")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.ProjectionSample.CreateMapAsync~Mapsui.Map")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.SvgSymbolSample.CreateMapAsync~Mapsui.Map")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Tests.Common.Maps.TilesSample.CreateMapAsync~Mapsui.Map")]
