+++ Tests/Mapsui.Tests.Common/Mapsui.Tests.Common.csproj
@@ -8,48 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove="Resources\Images\avion_silhouette.png" />
-    <None Remove="Resources\Images\checkered.png" />
-    <None Remove="Resources\Images\circle.png" />
-    <None Remove="Resources\Images\Ghostscript_Tiger.svg" />
-    <None Remove="Resources\Images\iconthatneedsoffset.png" />
-    <None Remove="Resources\Images\osm-liberty.png" />
-    <None Remove="Resources\Images\Pin.svg" />
-    <None Remove="Resources\SampleTiles\0_0_0.png" />
-    <None Remove="Resources\SampleTiles\1_0_0.png" />
-    <None Remove="Resources\SampleTiles\1_0_1.png" />
-    <None Remove="Resources\SampleTiles\1_1_0.png" />
-    <None Remove="Resources\SampleTiles\1_1_1.png" />
-    <None Remove="Resources\Shapefiles\test_file.dbf" />
-    <None Remove="Resources\Shapefiles\test_file.prj" />
-    <None Remove="Resources\Shapefiles\test_file.qmd" />
-    <None Remove="Resources\Shapefiles\test_file.shp" />
-    <None Remove="Resources\Shapefiles\test_file.shp.sidx" />
-    <None Remove="Resources\Shapefiles\test_file.shx" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Resources\Images\avion_silhouette.png" />
-    <EmbeddedResource Include="Resources\Images\checkered.png" />
-    <EmbeddedResource Include="Resources\Images\circle.png" />
-    <EmbeddedResource Include="Resources\Images\Ghostscript_Tiger.svg" />
-    <EmbeddedResource Include="Resources\Images\iconthatneedsoffset.png" />
-    <EmbeddedResource Include="Resources\Images\osm-liberty.png" />
-    <EmbeddedResource Include="Resources\Images\Pin.svg" />
-    <EmbeddedResource Include="Resources\SampleTiles\0_0_0.png" />
-    <EmbeddedResource Include="Resources\SampleTiles\1_0_0.png" />
-    <EmbeddedResource Include="Resources\SampleTiles\1_0_1.png" />
-    <EmbeddedResource Include="Resources\SampleTiles\1_1_0.png" />
-    <EmbeddedResource Include="Resources\SampleTiles\1_1_1.png" />
-    <EmbeddedResource Include="Resources\Shapefiles\test_file.dbf" />
-    <EmbeddedResource Include="Resources\Shapefiles\test_file.prj" />
-    <EmbeddedResource Include="Resources\Shapefiles\test_file.qmd" />
-    <EmbeddedResource Include="Resources\Shapefiles\test_file.shp" />
-    <EmbeddedResource Include="Resources\Shapefiles\test_file.shp.sidx" />
-    <EmbeddedResource Include="Resources\Shapefiles\test_file.shx" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\..\Mapsui.Tiling\Mapsui.Tiling.csproj" />
     <ProjectReference Include="..\..\Mapsui\Mapsui.csproj" />
     <ProjectReference Include="..\..\Samples\Mapsui.Samples.Common\Mapsui.Samples.Common.csproj" />
